Adding Jupiter and Saturn atmospheres
Lowering the '0 altitude' point for the remaining two gas giants to 1000
Bar, very crush depth.

Created a new pressure and temp curve for Jupiter based on real data;
copying to Saturn as well as a decent fill in for now. In Saturn's
atmosphere you would make it slightly deeper before being crushed to
death.