Added Biomes to Earth/Moon, Changed Moon Science
* Added completely new (and more accurate) biome map for Earth that has
11 total biomes (3 more than before). Decided not to change science
values around, but could lower the Low Space Values to 0.9 and High
Space to 1.25 which should make things closer to where they were before
(but increases the whack-a-mole)
* Added brand new Moon biome map that names the major features of the
Moon. This includes names for all of the Lunar Seas as well as the 10
largest craters. Total biomes are now 27 instead of 8.
* Lowered the Low Space Science Multiplier from 3.5 to 2 and the High
Space Multiplier from 3 to 2.5. This will still allow for more science
than before, but will limit some of the Whack-a-Mole aspects of it. We
can possibly go down to 1.5 for Low Space and 1.25 for High Space to get
closer to current Science Levels.