Fix PQS deactivate altitude for bodies
-This was obviously a placeholder value that was set to 167000 on all
bodies at one point. Since then it has been fixed on several bodies but
on Earth and Mars this remains above the atmosphere height which causes
the game to hang in orbit when you pass this altitude.

This has been fixed to bring it in line with all other bodies with an
atmosphere where it should deactivate below the atmo altitude