Turn off finalizeOrbit in Kopernicus configs (#285)
This causes problems due to the way that the KSP Orbit class is written
and the way that Kopernicus attempts to scribble over the meanMotion is
not internally consistent.  That would need to be patched and debugged
in stock fairly extensively.

The intended behavior also makes it difficult to simulate the Keplerian
motion of planets/satellites in RSS since the referenceBody.gravParameter
is incorrect and needs to be fixed by "reverse engineering" the effective
mu from the meanMotion on the celestial.  This adds a lot of headache.

Even apart from those issues, the buginess remains and causes issues
with Keplerian motion, and creates issues that have been seen with
detecting SOI encounters in the Earth-Moon system.  Attempting to fix
those issues by patching KSP for the meanMotion problems was not
successful either, which suggests some of the other tweaking that
finalizeOrbit is doing may be buggy as well.

https://github.com/KSPModdingLibs/KSPCommunityFixes/issues/141
https://github.com/Kopernicus/Kopernicus/issues/584

Merging this may not be the right way to fix this though since when I
tried disabling this value on a save, the Moon jumped by about 90
degrees in its orbit and a ManeuverNode which was on an encounter was
now no longer close.

This change may be save-breaking.